Decoupled logic of attachments from MessageInputController to AttachmentsController
