+++ library/src/main/java/com/getstream/sdk/chat/adapter/MediaAttachmentSelectedAdapter.kt
@@ -80,7 +80,8 @@ class MediaAttachmentSelectedAdapter(
             val cornerRadius =
                 itemView.context.resources.getDimensionPixelSize(R.dimen.stream_input_upload_media_radius)
             binding.ivMedia.setShape(
-                itemView.context, DrawableBuilder()
+                itemView.context,
+                DrawableBuilder()
                     .rectangle()
                     .solidColor(Color.BLACK)
                     .cornerRadii(cornerRadius, cornerRadius, cornerRadius, cornerRadius)
